Refactor purpose-made routes into separate modules
